% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPhysioExperiment-class.R
\name{alignment}
\alias{alignment}
\alias{alignment,MultiPhysioExperiment-method}
\alias{alignment<-}
\alias{alignment<-,MultiPhysioExperiment-method}
\title{Get or set temporal alignment metadata}
\usage{
alignment(x)

\S4method{alignment}{MultiPhysioExperiment}(x)

alignment(x) <- value

\S4method{alignment}{MultiPhysioExperiment}(x) <- value
}
\arguments{
\item{x}{A \code{MultiPhysioExperiment} object.}

\item{value}{A \code{\link[S4Vectors]{DataFrame}} with alignment metadata.}
}
\value{
A \code{\link[S4Vectors]{DataFrame}}.
}
\description{
Get or set temporal alignment metadata
}
\examples{
eeg_data <- matrix(rnorm(500 * 2), nrow = 500, ncol = 2)
pe_eeg <- PhysioExperiment(
  assays = list(raw = eeg_data),
  samplingRate = 250
)
mpe <- MultiPhysioExperiment(EEG = pe_eeg)
alignment(mpe)
}
\seealso{
\code{\link[=alignSignals]{alignSignals()}}, \code{\link[=experiments]{experiments()}}, \code{\link[=samplingRates]{samplingRates()}}
}
