% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPhysioExperiment-class.R
\name{.build_default_alignment}
\alias{.build_default_alignment}
\title{Build default alignment DataFrame from experiments}
\usage{
.build_default_alignment(exps)
}
\arguments{
\item{exps}{Named list of PhysioExperiment objects.}
}
\value{
A DataFrame with one row per modality.
}
\description{
Build default alignment DataFrame from experiments
}
\keyword{internal}
