% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPhysioExperiment-class.R
\name{nModalities}
\alias{nModalities}
\alias{nModalities,MultiPhysioExperiment-method}
\title{Get number of modalities}
\usage{
nModalities(x)

\S4method{nModalities}{MultiPhysioExperiment}(x)
}
\arguments{
\item{x}{A \code{MultiPhysioExperiment} object.}
}
\value{
Integer scalar: the number of modalities stored in the object.
}
\description{
Get number of modalities
}
\examples{
eeg_data <- matrix(rnorm(500 * 2), nrow = 500, ncol = 2)
pe_eeg <- PhysioExperiment(
  assays = list(raw = eeg_data),
  samplingRate = 250
)
mpe <- MultiPhysioExperiment(EEG = pe_eeg)
nModalities(mpe)
}
\seealso{
\code{\link[=modalities]{modalities()}}, \code{\link[=experiments]{experiments()}}
}
