% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{MultiPhysioExperiment}
\alias{experiments}
\alias{experiments<-}
\alias{modalities}
\alias{samplingRates}
\alias{nModalities}
\alias{alignment}
\alias{alignment<-}
\alias{streams}
\alias{commonClock}
\alias{timeWindow}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{PhysioExperiment}{\code{\link[PhysioExperiment:multiphysio-legacy-accessors]{alignment}}, \code{\link[PhysioExperiment:multiphysio-legacy-accessors]{alignment<-}}, \code{\link[PhysioExperiment]{commonClock}}, \code{\link[PhysioExperiment:multiphysio-legacy-accessors]{experiments}}, \code{\link[PhysioExperiment:multiphysio-legacy-accessors]{experiments<-}}, \code{\link[PhysioExperiment:multiphysio-legacy-accessors]{modalities}}, \code{\link[PhysioExperiment]{MultiPhysioExperiment}}, \code{\link[PhysioExperiment:multiphysio-legacy-accessors]{nModalities}}, \code{\link[PhysioExperiment:multiphysio-legacy-accessors]{samplingRates}}, \code{\link[PhysioExperiment]{streams}}, \code{\link[PhysioExperiment]{timeWindow}}}
}}

