% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPhysioExperiment-class.R
\name{[[,MultiPhysioExperiment,ANY,ANY-method}
\alias{[[,MultiPhysioExperiment,ANY,ANY-method}
\title{Extract a single modality from a MultiPhysioExperiment}
\usage{
\S4method{[[}{MultiPhysioExperiment,ANY,ANY}(x, i, j, ...)
}
\arguments{
\item{x}{A \code{MultiPhysioExperiment} object.}

\item{i}{Modality name (character) or index (integer).}

\item{j}{Not used.}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{PhysioExperiment} object.
}
\description{
Extract a single modality from a MultiPhysioExperiment
}
\examples{
eeg_data <- matrix(rnorm(500 * 2), nrow = 500, ncol = 2)
pe_eeg <- PhysioExperiment(
  assays = list(raw = eeg_data),
  samplingRate = 250
)
mpe <- MultiPhysioExperiment(EEG = pe_eeg)
mpe[["EEG"]]
}
